+++ packages/boring-table-core/src/plugins/index.ts
@@ -1,6 +1,7 @@
+export * from './RowSelection';
+export { BoringPlugin } from './base';
 export type { IBoringPlugin } from './base';
 export * from './change';
-export * from './check';
 export * from './filter';
 export * from './hidden';
 export * from './pagination';