+++ packages/boring-table-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # boring-table
 
+## 0.0.7
+
+### Patch Changes
+
+- add custom body
+  add pagination
+  add filter
+
+  fix types
+
+## 0.0.6
+
+### Patch Changes
+
+- update logo path
+
 ## 0.0.5
 
 ### Patch Changes